double->BigDecimal

+++ src/main/java/main/model/Goods.java
@@ -1,11 +1,11 @@
 package main.model;
 import javax.persistence.Entity;
-
+import java.math.BigDecimal;
 
 public interface Goods {
 
 	String getName();
-	Double getPrice();
+	BigDecimal getPrice();
 	String getType();
 	String getPackage();
 	Long getId();
